Revert "Rework version"
