--- new_src/utils/clipboard.py
@@ -1,105 +1,0 @@
-import win32clipboard
-from io import BytesIO
-from PIL import Image
-import win32gui
-import os
-import win32ui
-from ctypes import windll
-
-#####################################################
-#                                                   #
-#             Clipboard stuff                       #          
-#                                                   #
-######################################################
-
-def copy_im_to_clipboard(image):
-    bio = BytesIO()
-    image.save(bio, 'BMP')
-    data = bio.getvalue()[14:] # removing some headers
-    bio.close()
-    send_to_clipboard(win32clipboard.CF_DIB, data)
-
-
-def send_to_clipboard(clip_type, data):
-    if clip_type == "text":
-        win32clipboard.OpenClipboard()
-        win32clipboard.EmptyClipboard()
-        win32clipboard.SetClipboardText(data)
-        win32clipboard.CloseClipboard()
-    else:
-        win32clipboard.OpenClipboard()
-        win32clipboard.EmptyClipboard()
-        win32clipboard.SetClipboardData(clip_type, data)
-        win32clipboard.CloseClipboard()
-
-def get_clipboard_data():
-    try:
-        win32clipboard.OpenClipboard()
-        data = win32clipboard.GetClipboardData()
-        win32clipboard.CloseClipboard()
-        print("HELLO?")
-        return data
-    except TypeError as err:
-        print("BROKEN")
-        return "Invalid Clipboard Data"
-
-
-## Image to Bytes
-def file_to_bytes(filepath):
-    ## Take image into bytes and onto clipboard
-    image = Image.open(filepath)
-    output = BytesIO()
-    image.convert("RGB").save(output, "BMP")
-    data = output.getvalue()[14:]
-    output.close()
-    
-    ### Sending to Clipboard
-    send_to_clipboard(win32clipboard.CF_DIB, data)
-    ### Deleting Temp File
-    os.remove(filepath)
-    print("Temp Image Deleted")
-
-###screenshot window without bringing it to foreground 
-def screenshot_window(capture_type, window_title=None, clipboard=False, save_location=None):
-    hwnd = win32gui.FindWindow(None, window_title)
-    try:
-        left, top, right, bot = win32gui.GetClientRect(hwnd)
-        #left, top, right, bot = win32gui.GetWindowRect(hwnd)
-        w = right - left
-        h = bot - top
-        
-        hwndDC = win32gui.GetWindowDC(hwnd)
-        mfcDC  = win32ui.CreateDCFromHandle(hwndDC)
-        saveDC = mfcDC.CreateCompatibleDC()
-        
-        saveBitMap = win32ui.CreateBitmap()
-        saveBitMap.CreateCompatibleBitmap(mfcDC, w, h)
-        saveDC.SelectObject(saveBitMap)
-        
-        # Change the line below depending on whether you want the whole window
-        # or just the client area as shown above. 
-                              # 1, 2, 3 all give different results   ( 3 seems to work for everything)
-        result = windll.user32.PrintWindow(hwnd, saveDC.GetSafeHdc(), capture_type)
-        bmpinfo = saveBitMap.GetInfo()
-        bmpstr = saveBitMap.GetBitmapBits(True)
-        
-        im = Image.frombuffer(
-            'RGB',
-            (bmpinfo['bmWidth'], bmpinfo['bmHeight']),
-            bmpstr, 'raw', 'BGRX', 0, 1)
-        
-        win32gui.DeleteObject(saveBitMap.GetHandle())
-        saveDC.DeleteDC()
-        mfcDC.DeleteDC()
-        win32gui.ReleaseDC(hwnd, hwndDC)
-        
-        if result == 1:
-            #PrintWindow Succeeded
-            if clipboard == True:
-                copy_im_to_clipboard(im)
-                print("Copied to Clipboard")
-            elif clipboard == False:
-                im.save(save_location+".png")
-                print("Saved to Folder")
-    except Exception as e:
-        print("error screenshot" + e )
