--- new_src/utils/util.py
@@ -1,213 +1,0 @@
-import os
-import subprocess
-from ctypes import windll
-import win32gui
-import pyautogui
-
-def windowsSettings(choice=False):
-    
-    settings ={
-        "SYSTEM:  Display": 'ms-settings:display',
-        "SYSTEM:  Advanced Display": 'ms-settings:display-advanced',
-        "SYSTEM:  Night Light": 'ms-settings:nightlight',
-        "SYSTEM:  Sound": 'ms-settings:sound',
-        "SYSTEM:  Manage Sound Devices": 'ms-settings:sound-devices',
-        "SYSTEM:  Manage App/Device Volume": 'ms-settings:apps-volume',
-        "SYSTEM:  App Volume & Device Preferences": 'ms-settings:apps-volume',
-        "SYSTEM:  Notifcations & Actions": 'ms-settings:notifications',
-        "SYSTEM:  Power & Sleep": 'ms-settings:powersleep',
-        "SYSTEM:  Battery": 'ms-settings:batterysaver',
-        "SYSTEM:  Battery Usage Details": 'ms-settings:batterysaver-usagedetails',
-        "SYSTEM:  Default Save Locations": 'ms-settings:savelocations',
-        "SYSTEM:  Multi-Tasking": 'ms-settings:multitasking',
-        "SYSTEM:  Sign-in Options": 'ms-settings:signinoptions',     ## this was 'ACCOUNTS'
-        "SYSTEM:  Date & Time": 'ms-settings:dateandtime',             ## this was 'TIME & LANGUAGE*
-        "SYSTEM:  Time Region": 'ms-settings:regionformatting',        ## this was 'TIME & LANGUAGE*
-        "SYSTEM:  Settings Home Page": 'ms-settings:',
-        "NETWORK:  Ethernet": 'ms-settings:network-ethernet',
-        "NETWORK:  Wi-Fi": 'ms-settings:network-wifi',
-        "PERSONALIZATION:  Background": 'ms-settings:personalization-background',
-        "PERSONALIZATION:  Colors": 'ms-settings:personalization-colors',
-        "PERSONALIZATION:  Lock Screen": 'ms-settings:lockscreen',
-        "PERSONALIZATION:  Themes": 'ms-settings:themes',
-        "PERSONALIZATION:  Start Folders": 'ms-settings:personalization-start-places',
-        "APPS:  Apps & Features": 'ms-settings:appsfeatures',
-        "APPS:  Manage Startup Apps": 'ms-settings:startupapps',
-        "APPS:  Manage Default Apps": 'ms-settings:defaultapps',
-        "APPS:  Manage Optional Features": 'ms-settings:optionalfeatures',
-        "GAMING:  Game Bar": 'ms-settings:gaming-gamebar',
-        "GAMING:  Game DVR": 'ms-settings:gaming-gamedvr',
-        "GAMING:  Game Mode": 'ms-settings:gaming-gamemode',
-        "GAMING:  XBOX Networking": 'ms-settings:gaming-xboxnetworking',
-        "PRIVACY:  Activity History": 'ms-settings:privacy-activityhistory',
-        "PRIVACY:  Webcam": 'ms-settings:privacy-webcam',
-        "PRIVACY:  Microphone": 'ms-settings:privacy-microphone',
-        "PRIVACY:  Background Apps": 'ms-settings:privacy-backgroundapps',
-        r"UPDATE & SECURITY:  Windows Update": 'ms-settings:windowsupdate',
-        r"UPDATE & SECURITY:  Windows Recovery": 'ms-settings:recovery',
-        r"UPDATE & SECURITY:  Update history": 'ms-settings:windowsupdate-history',
-        r"UPDATE & SECURITY:  Restart Options": 'ms-settings:windowsupdate-restartoptions',
-        r"UPDATE & SECURITY:  Delivery Optimization": 'ms-settings:delivery-optimization',
-        r"UPDATE & SECURITY:  Windows Security": 'ms-settings:windowsdefender',
-        r"UPDATE & SECURITY:  Windows Defender": 'windowsdefender:',
-        r"UPDATE & SECURITY:  For Developers": 'ms-settings:developers',
-    }
-    if not choice:
-        return list(settings.keys())
-    else:
-        os.system(f'explorer "{settings[choice]}"')
-
-def runCommandLine(command):
-    systemencoding = windll.kernel32.GetConsoleOutputCP()
-    systemencoding= f"cp{systemencoding}"
-    output = subprocess.run(command, stdout=subprocess.PIPE, shell=True)
-    result = str(output.stdout.decode(systemencoding))
-    return result
-
-def get_windows():
-    results = []
-    def winEnumHandler(hwnd, ctx):
-        if win32gui.IsWindowVisible(hwnd):
-            if win32gui.GetWindowText(hwnd):
-    
-                results.append(win32gui.GetWindowText(hwnd))
-                
-    win32gui.EnumWindows(winEnumHandler, None)
-    return results
-
-def AdvancedMouseFunction(x, y, delay, look):
-    if look == 0:
-        look = None
-    if look == "None":
-        try:
-            pyautogui.moveTo(x, y, delay)
-        except:
-            pass
-    elif look == "Start slow, end fast":
-        try:
-            pyautogui.moveTo(x, y, delay, pyautogui.easeInQuad)
-        except:
-            pass
-    elif look == "Start fast, end slow":
-        try:
-            pyautogui.moveTo(x, y, delay, pyautogui.easeOutQuad)
-        except:
-            pass
-    elif look == "Start and end fast, slow in middle":
-        try:
-            pyautogui.moveTo(x, y, delay, pyautogui.easeInOutQuad)
-        except:
-            pass
-    elif look == "bounce at the end":
-        try:
-            pyautogui.moveTo(x, y, delay, pyautogui.easeInBounce)
-        except:
-            pass
-    elif look == "rubber band at the end":
-        try:
-            pyautogui.moveTo(x, y, delay, pyautogui.easeInElastic)
-        except:
-            pass
-
-import json
-
-def AudioDeviceCmdlets(command, output=True):
-    ### add in another coinitilize???  
-    
-    systemencoding = windll.kernel32.GetConsoleOutputCP()
-    systemencoding= f"cp{systemencoding}"
-    process = subprocess.Popen(["powershell", "-Command", "Import-Module .\AudioDeviceCmdlets.dll;", command],stdout=subprocess.PIPE, shell=True, encoding=systemencoding)
-    proc_stdout = process.communicate()[0]
-    if output:
-        proc_stdout = proc_stdout[proc_stdout.index("["):-1]
-        print(proc_stdout)
-        return json.loads(proc_stdout) 
-
-
-import pyttsx3
-import sounddevice as sd
-
-
-def getAllVoices():
-    engine = pyttsx3.init()
-    return engine.getProperty("voices")
-
-def getAllOutput_TTS2():
-    audio_dict = {}
-    for outputDevice in sd.query_hostapis(0)['devices']:
-        if sd.query_devices(outputDevice)['max_output_channels'] > 0:
-            audio_dict[sd.query_devices(device=outputDevice)['name']] = sd.query_devices(device=outputDevice)['default_samplerate']
-    return audio_dict
-
-
-import ctypes
-
-import numpy as np
-import pyautogui
-import win32gui
-from PIL import Image
-
-
-def get_windows():
-    results = []
-    def winEnumHandler(hwnd, ctx):
-        if win32gui.IsWindowVisible(hwnd):
-            if win32gui.GetWindowText(hwnd):
-    
-                results.append(win32gui.GetWindowText(hwnd))
-                
-    win32gui.EnumWindows(winEnumHandler, None)
-    return results
-
-def winextra(action):
-    
-  # if action == "Keep Active, Minimize All (Toggle)":
-  #     pyautogui.hotkey('win', 'home')
-  #    
-  # if action == "Minimize All (Toggle)":
-  #     pyautogui.hotkey('win', 'd')
-#         
-  #  if action =="Clipboard History":
-  #      pyautogui.hotkey('win', 'v')
-        
-    if action == "Emoji":
-        pyautogui.hotkey('win', '.')
-        
-    if action == "Keyboard":
-        pyautogui.hotkey('win', 'ctrl', "o")
-
-def get_key_state(key):
-    hllDll = ctypes.WinDLL ("User32.dll")
-    if key == "NUM LOCK":
-        return hllDll.GetKeyState(0x90)
-    if key == "CAPS LOCK":
-        return hllDll.GetKeyState(0x14)
-
-def move_win_button(direction):
-    check = get_key_state('NUM LOCK')
-    if not check:
-        pyautogui.hotkey('win', 'shift', direction)
-    elif check:
-        pyautogui.press('numlock')
-        pyautogui.hotkey('win', 'shift', direction)
-        pyautogui.press('numlock')
-
-def resize_image(im, height, width):
-    if im.size[0] == height and im.size[1] == width: 
-        return im
-    else:
-        print("We are actually resizing")
-        size = (height,width)
-        im.load()
-        bands = list(im.split())
-        a = np.asarray(bands[-1])
-        a.flags.writeable = True
-        a[a != 0] = 1
-        bands[-1] = Image.fromarray(a)
-        bands = [b.resize(size, Image.LINEAR) for b in bands]
-        a = np.asarray(bands[-1])
-        a.flags.writeable = True
-        a[a != 0] = 255
-        bands[-1] = Image.fromarray(a)
-        im = Image.merge('RGBA', bands)
-        return im
