+++ src/requirements.txt
@@ -1,0 +1,45 @@
+altgraph==0.17.2
+audio2numpy==0.1.2
+certifi==2021.10.8
+cffi==1.15.0
+charset-normalizer==2.0.9
+comtypes==1.1.10
+ffmpeg==1.4
+future==0.18.2
+idna==3.3
+MouseInfo==0.1.3
+mss==6.1.0
+numpy==1.22.1
+pefile==2021.9.3
+Pillow==9.0.0
+psutil==5.9.0
+PyAutoGUI==0.9.53
+https://github.com/AndreMiras/pycaw/archive/refs/tags/v20190904.zip
+pycparser==2.21
+pyee==8.2.2
+PyGetWindow==0.0.9
+PyGetWindowMP @ https://github.com/Kalmat/PyGetWindow/archive/refs/heads/master.zip
+pyinstaller==4.7
+pyinstaller-hooks-contrib==2021.4
+PyMsgBox==1.0.9
+pyperclip==1.8.2
+pyping==0.0.6
+pypiwin32==223
+PyRect==0.1.4
+PyScreeze==0.1.28
+python-dateutil==2.8.2
+pyttsx3==2.90
+pytweening==1.0.4
+pyvda==0.2.4
+pywin32==302
+pywin32-ctypes==0.2.0
+requests==2.27.0
+schedule==1.1.0
+scipy==1.7.3
+screeninfo==0.8
+six==1.16.0
+sounddevice==0.4.4
+TouchPortal-API==1.6
+urllib3==1.26.7
+winotify==1.1.0
+wintoast==0.2
